change category atctivate

+++ src/core/src/modules/category/application/usecases/delete-category/delete-category.usecase.ts
@@ -1,8 +1,9 @@
-import { Either, right } from "@/modules/@shared/logic";
+import { Either, left, right } from "@/modules/@shared/logic";
 import {  EventEmitterInterface } from "@/modules/@shared/events";
 import { CategoryDeletedEvent } from "./category-deleted.event";
 import { CategoryRepositoryInterface } from "@/modules/category/domain/repositories";
 import { DeleteCategoryUsecaseInterface } from "@/modules/category/domain/usecases";
+import { CategoryNotFoundError } from "../_errors";
 
 export class DeleteCategoryUsecase implements DeleteCategoryUsecaseInterface {
 
@@ -12,6 +13,9 @@ export class DeleteCategoryUsecase implements DeleteCategoryUsecaseInterface {
     ){}
 
     async execute({ categoryId }: DeleteCategoryUsecaseInterface.InputDto): Promise<DeleteCategoryUsecaseInterface.OutputDto> {
+
+        const categoryEntity = await this.categoryRepository.findById(categoryId)
+        if(!categoryEntity) return left([ new CategoryNotFoundError() ])
 
         await this.categoryRepository.delete(categoryId)
